+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## 1.1.0
 
- - `TreeMaker` link to parent in `TreeItem`
+ - `TreeMaker` 
+   - link to parent in `TreeItem`
+   - 
  - `CollectionDiff` Input collections can be iterable. (Originally only array)
+ - `ProcI`
+   - method `each`
 
 ## 1.0.0 First published version